gitk: fix history window panes position
When the history window panes in are resized
by moving either of the two sashes and then
the gitk window is vertically resized,
the sashes fall back into their previous
position without respecting the users wish
for resizing.
Save the sash position when the sashes are
moved to make them keep their position when
the window is resized afterwards.

When the gitk window is opened and maximized
on a screen, then closed and opened on a
screen smaller than the previously used one,
the author pane and time pane of the history
window only are a few pixels wide and their
contents are barely visible.
Widen the two panes on start of gitk to a
reasonable fixed size that shows a good amount
of text of authors and time.

Signed-off-by: Tobias Boesch <tobias.boesch@miele.com>